requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
python-multipart
# (//TODO: fastapi-csrf-protect after integrating CSRF protection)
